requirements.txt
pandas==2.2.2
openpyxl==3.1.2
pydantic==2.7.1
sqlalchemy==2.0.30
google-genai>=1.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
